requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
black==24.4.0
click==8.1.7
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
isort==5.13.2
joblib==1.3.2
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pyasn1==0.6.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
threadpoolctl==3.4.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
